Add new InfiniteScroll storybook doc
